packages/agentkit-eval/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentkit-eval"
version = "0.1.0"
description = "Declarative evaluation harness for AgentKit agents (suite.yaml -> pass-rate / latency / cost report)."
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "agentkit>=0.1.0",
    "pyyaml>=6.0",
    "rich>=13.7",
    "typer>=0.12",
]

[project.scripts]
agentkit-eval = "agentkit_eval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentkit_eval"]
